broken.line {segmented}	R Documentation

Fitted values for segmented relationships

Description

Given a segmented model (typically returned by a segmented method), broken.line computes the fitted values (and relevant standard errors) for the specified ‘segmented’ relationship.

Usage

broken.line(ogg, term = NULL, link = TRUE, interc=TRUE, se.fit=TRUE, isV=FALSE, 
  .vcov=NULL, .coef=NULL, ...)

Arguments

ogg
A fitted object of class segmented (returned by any segmented method).

term
Three options. i) A named list (whose name should be one of the segmented covariates in the model ogg) including the covariate values for which segmented predictions should be computed; ii) a character meaning the name of any segmented covariate in the model (and predictions corresponding to the observed covariate values are returned); iii) It can be NULL if the model includes a single segmented covariate (and predictions corresponding to the observed covariate values are returned).

link
Should the predictions be computed on the scale of the link function if ogg is a segmented glm fit? Default to TRUE.

interc
Should the model intercept be added? (provided it exists).

se.fit
If TRUE also standard errors for predictions are returned.

isV
A couple of logicals indicating if the segmented terms (x-ψ)_+ and I(x>ψ) in the model matrix should be replaced by their smoothed counterparts when computing the standard errors. If a single logical is provided, it is applied to both terms.

.vcov
Optional. The full covariance matrix of estimates. If NULL (and se.fit=TRUE), the matrix is computed internally via vcov.segmented().

.coef
The regression parameter estimates. If unspecified (i.e. NULL), it is computed internally by coef(ogg).

...
Additional arguments to be passed on to vcov.segmented() when computing the standard errors for the predictions, namely is, var.diff, p.df. See summary.segmented and vcov.segmented.

Details

If term=NULL or term is a valid segmented covariate name, predictions for that segmented variable are the relevant fitted values from the model. If term is a (correctly named) list with numerical values, predictions corresponding to such specified values are computed. If link=FALSE and ogg inherits from the class "glm", predictions and possible standard errors are returned on the response scale. The standard errors come from the Delta method. Argument link is ignored whether ogg does not inherit from the class "glm".

Value

A list having one component if (if se.fit=FALSE), and two components (if se.fit=TRUE) list representing predictions and standard errors for the segmented covariate values.

Note

This function was written when there was not predict.segmented (which is more general).

Author(s)

Vito M. R. Muggeo

See Also

segmented, predict.segmented, plot.segmented, vcov.segmented

Examples

set.seed(1234)
z<-runif(100)
y<-rpois(100,exp(2+1.8*pmax(z-.6,0)))
o<-glm(y~z,family=poisson)
o.seg<-segmented(o,seg.Z=~z)
## Not run: plot(z,y)
## Not run: points(z,broken.line(o.seg,link=FALSE)$fit,col=2) #ok, but use plot.segmented()!
    

[Package segmented version 2.1-0 Index]