confint.segmented.lme {segmented}	R Documentation

Confidence intervals in segmented mixed models

Description

Computes confidence intervals for all regression parameters, including the the breakpoint, in a fitted ‘segmented mixed’ model.

Usage

## S3 method for class 'segmented.lme'
confint(object, parm, level = 0.95, obj.boot, ...)

Arguments

object
A fit object returned by segmented.lme.

parm
A vector of numbers indicating which parameters should be considered. If missing all parameters.

level
The confidence level.

obj.boot
The possible list including the bootstrap distributions of the regression coefficients. Such list is returned by vcov.segmented.lme(.., ret.b=TRUE)

...
if obj.boot is missing and bootstrap CIs are requested, additional optional arguments, such as B, seed, and it.max.b, to be used in computations of the boot distributions.

Details

If obj.boot is provided or ... includes the argument B>0, confidence intervals are computed by exploiting the bootstrap distributions.

Value

A matrix (or a list of matrices if bootstrap ci are requested) including the confidence intervals for the model parameters.

Warning

All the functions for segmented mixed models (*.segmented.lme) are still at an experimental stage

Author(s)

Vito Muggeo

See Also

vcov.segmented.lme

Examples

## Not run: 
confint(os) #asymptotic CI

confint(os, B=50) #boot CIs

#it is possible to obtain the boot distribution beforehand
ob <-vcov(os, B=50, ret.b=TRUE)
confint(os, obj.boot=ob) #boot CI

## End(Not run)

[Package segmented version 2.1-0 Index]