down {segmented}	R Documentation

Down syndrome in babies

Description

The down data frame has 30 rows and 3 columns. Variable cases means the number of babies with Down syndrome out of total number of births births for mothers with mean age age.

Usage

data(down)

Format

A data frame with 30 observations on the following 3 variables.

age

the mothers' mean age.

births

count of total births.

cases

count of babies with Down syndrome.

Source

Davison, A.C. and Hinkley, D. V. (1997) Bootstrap Methods and their Application. Cambridge University Press.

References

Geyer, C. J. (1991) Constrained maximum likelihood exemplified by isotonic convex logistic regression. Journal of the American Statistical Association 86, 717–724.

Examples

data(down)
    

[Package segmented version 2.1-0 Index]