fitted.segmented.lme {segmented}	R Documentation

Fitted values for segmented mixed fits

Description

Computes fitted values at different levels of nesting for segmented mixed objects

Usage

## S3 method for class 'segmented.lme'
fitted(object, level = 1, sort=TRUE, ...)

Arguments

object
Object of class "segmented.lme"

level
the level to be considered. Currently only levels 0 or 1 are allowed.

sort
If TRUE, the fitted values are sorted by the names of the 'id' levels.

...
Ignored

Details

Currently it works only if level=1

Value

A numeric object including the fitted values at the specified level of nesting.

Warning

All the functions for segmented mixed models (*.segmented.lme) are still at an experimental stage

Author(s)

Vito Muggeo

See Also

summary.segmented.lme

[Package segmented version 2.1-0 Index]