globTempAnom {segmented}	R Documentation

Global temperature anomalies 1850-2023

Description

The globTempAnom data frame includes the global surface temperature anomalies from 1850 to 2023.

Usage

data(globTempAnom)

Format

The included variables are (clearly).

Year

the calendar year.

Anomaly

the temperature anomalies computed as differences of the annual (average) measurement with respect to the 20th century average (1901-2000).

Details

Data refer to averages measurements referring to land and ocean surface of Northern and Southern hemisphere.

Source

https://www.ncei.noaa.gov/access/monitoring/global-temperature-anomalies/anomalies

References

There are several references using such dataset, e.g.

Cahill, N., Rahmstorf, S., and Parnell, A. C. (2015). Change points of global temperature. Environmental Research Letters, 10: 1-6.

Examples

data(globTempAnom)
    

[Package segmented version 2.1-0 Index]