intercept {segmented}	R Documentation

Intercept estimates from segmented relationships

Description

Computes the intercepts of each ‘segmented’ relationship in the fitted model.

Usage

intercept(ogg, parm, rev.sgn = FALSE, var.diff=FALSE,
    .vcov=NULL, .coef=NULL, digits = max(4, getOption("digits") - 2),...)

Arguments

ogg
an object of class "segmented", returned by any segmented method.

parm
the segmented variable whose intercepts have to be computed. If missing all the segmented variables in the model are considered.

rev.sgn
vector of logicals. The length should be equal to the length of parm, but it is recycled otherwise. When TRUE it is assumed that the current parm is ‘minus’ the actual segmented variable, therefore the order is reversed before printing. This is useful when a null-constraint has been set on the last slope.

var.diff
Currently ignored as only point estimates are computed.

.vcov
The full covariance matrix of estimates. If unspecified (i.e. NULL), the covariance matrix is computed internally by vcov(ogg).

.coef
The regression parameter estimates. If unspecified (i.e. NULL), it is computed internally by coef(ogg).

digits
controls number of digits in the returned output.

...
Further arguments to be passed on to vcov.segmented, such as var.diff and is. See Details in vcov.segmented.

Details

A broken-line relationship means that a regression equation exists in the intervals 'min(x) to psi1', 'psi1 to psi2', and so on. intercept computes point estimates of the intercepts of the different regression equations for each segmented relationship in the fitted model.

Value

intercept returns a list of one-column matrices. Each matrix represents a segmented relationship.

Author(s)

Vito M. R. Muggeo, vito.muggeo@unipa.it

See Also

See also slope to compute the slopes of the different regression equations for each segmented relationship in the fitted model.

Examples

## see ?slope
## Not run: 
intercept(out.seg)

## End(Not run)

[Package segmented version 2.1-0 Index]