lines.stepmented {segmented}	R Documentation

Bars for interval estimate of the breakpoints

Description

Draws bars relevant to breakpoint estimates (point estimate and confidence limits) on the current device

Usage

## S3 method for class 'stepmented'
lines(x, term, bottom = TRUE, shift=FALSE, conf.level = 0.95, k = 50, 
    pch = 18, .vcov=NULL, .coef=NULL, ...)

Arguments

x
an object of class stepmented.

term
the stepmented variable of the breakpoints being drawn. It may be unspecified when there is a single stepmented variable.

bottom
logical, indicating if the bars should be plotted at the bottom (TRUE) or at the top (FALSE).

shift
logical, indicating if the bars should be ‘shifted’ on the y-axis before plotting. Useful for multiple breakpoints with overlapped confidence intervals.

conf.level
the confidence level of the confidence intervals for the breakpoints.

k
a positive integer regulating the vertical position of the drawn bars. See Details.

pch
either an integer specifying a symbol or a single character to be used in plotting the point estimates of the breakpoints. See points.

.vcov
The full covariance matrix of estimates. If unspecified (i.e. NULL), the covariance matrix is computed internally by vcov(x).

.coef
The regression parameter estimates. If unspecified (i.e. NULL), it is computed internally by coef(x).

...
further arguments passed to arrows, for instance ‘col’ that can be a vector.

Details

lines.stepmented simply draws on the current device the point estimates and relevant confidence limits of the estimated breakpoints from a "stepmented" object. The y coordinates where the bars are drawn is computed as usr[3]+h if bottom=TRUE or usr[4]-h when bottom=FALSE, where h=(usr[4]-usr[3])/abs(k) and usr are the extremes of the user coordinates of the plotting region. Therefore for larger values of k the bars are plotted on the edges.

See Also

plot.stepmented to plot the fitted stepmented lines

Examples

## See ?plot.stepmented

[Package segmented version 2.1-0 Index]