predict.stepmented {segmented}	R Documentation

Predict method for stepmented model fits

Description

Returns predictions and optionally associated quantities (standard errors or confidence intervals) from a fitted stepmented model object.

Usage

## S3 method for class 'stepmented'
predict(object, newdata, se.fit=FALSE, interval=c("none","confidence", "prediction"), 
            type = c("link", "response"), na.action=na.omit, level=0.95, .coef=NULL, 
            .vcov=NULL, apprx.fit=c("none","cdf"), apprx.se=c("cdf","none"), ...)

Arguments

object
a fitted stepmented model coming from stepmented.lm or stepmented.glm.

newdata
An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.

se.fit
Logical. Should the standard errors be returned?

interval
Which interval? See predict.lm

type
Predictions on the link or response scale? Only if object is a stepmented glm.

na.action
How to deal with missing data, if newdata include them.

level
The confidence level.

.coef
The regression parameter estimates. If unspecified (i.e. NULL), it is computed internally by coef().

.vcov
The estimate covariance matrix. If unspecified (i.e. NULL), it is computed internally by vcov.stepmented().

apprx.fit
The approximation of the (x>\hatψ) used to compute the predictions/fitted values of the piece-wise relationships.

apprx.se
The same abovementioned approximation to compute the standard error.

...
further arguments, for instance k to be passed to vcov.stepmented.

Details

Basically predict.stepmented builds the right design matrix accounting for breakpoint and passes it to predict.lm or predict.glm depending on the actual model fit object.

Value

predict.stepmented produces a vector of predictions with possibly associated standard errors or confidence intervals. See predict.lm, predict.glm, or predict.segmented.

Warning

For stepmented glm fits with offset obtained starting from the model glm(.., offset=..), predict.stepmented returns the fitted values without considering the offset.

Author(s)

Vito Muggeo

See Also

stepmented, plot.stepmented, predict.lm, predict.glm

Examples

n=10
x=seq(-3,3,l=n)
set.seed(1515)
y <- (x<0)*x/2 + 1 + rnorm(x,sd=0.15)
segm <- segmented(lm(y ~ x), ~ x, psi=0.5)
predict(segm,se.fit = TRUE)$se.fit

[Package segmented version 2.1-0 Index]