stagnant {segmented}	R Documentation

Stagnant band height data

Description

The stagnant data frame has 28 rows and 2 columns.

Usage

data(stagnant)

Format

A data frame with 28 observations on the following 2 variables.

x

log of flow rate in g/cm sec.

y

log of band height in cm

Details

Bacon and Watts report that such data were obtained by R.A. Cook during his investigation of the behaviour of stagnant surface layer height in a controlled flow of water.

Source

Bacon D.W., Watts D.G. (1971) Estimating the transistion between two intersecting straight lines. Biometrika 58: 525 – 534.

Originally from the PhD thesis by R.A. Cook

Examples

data(stagnant)
## plot(stagnant)

[Package segmented version 2.1-0 Index]