step.lm.fit {segmented}	R Documentation

Fitter Functions for stepmented Linear Models

Description

step.lm.fit is called by stepmented.lm to fit stepmented linear (gaussian) models. Likewise, step.glm.fit is called by stepmented.glm to fit generalized stepmented linear models. The step.*.fit.boot functions are employed to perform bootstrap restarting. These functions should usually not be used directly by the user.

Usage

step.lm.fit(y, x.lin, Xtrue, PSI, ww, offs, opz, return.all.sol=FALSE)  

step.lm.fit.boot(y, XREG, Z, PSI, w, offs, opz, n.boot=10, size.boot=NULL, 
        jt=FALSE, nonParam=TRUE, random=FALSE, break.boot=n.boot)                          

step.glm.fit(y, x.lin, Xtrue, PSI, ww, offs, opz, return.all.sol=FALSE)

step.glm.fit.boot(y, XREG, Z, PSI, w, offs, opz, n.boot=10, size.boot=NULL, 
        jt=FALSE, nonParam=TRUE, random=FALSE, break.boot=n.boot)

 

 

Arguments

y
vector of observations of length n.

x.lin, XREG
design matrix for standard linear terms.

Xtrue, Z
appropriate matrix including the stepmented variables whose breakpoints have to be estimated.

PSI
appropriate matrix including the starting values of the breakpoints to be estimated.

ww,w
possibe weights vector.

offs
possibe offset vector.

opz
a list including information useful for model fitting.

n.boot
the number of bootstrap samples employed in the bootstrap restart algorithm.

break.boot
Integer, less than n.boot. If break.boot consecutive bootstrap samples lead to the same objective function, the algorithm stops without performing all n.boot 'trials'. This can save computational time considerably.

size.boot
the size of the bootstrap resamples. If NULL (default), it is taken equal to the sample size. values smaller than the sample size are expected to increase perturbation in the bootstrap resamples.

jt
logical. If TRUE the values of the stepmented variable(s) are jittered before fitting the model to the bootstrap resamples.

nonParam
if TRUE nonparametric bootstrap (i.e. case-resampling) is used, otherwise residual-based.

random
if TRUE, when the algorithm fails to obtain a solution, random values are used as candidate values.

return.all.sol
if TRUE, when the algorithm fails to obtain a solution, the values visited by the algorithm with corresponding deviances are returned.

Details

The functions call iteratively lm.wfit (or glm.fit) with proper design matrix depending on XREG, Z and PSI. step.lm.fit.boot (and step.glm.fit.boot) implements the bootstrap restarting idea discussed in Wood (2001).

Value

A list of fit information.

Note

These functions should usually not be used directly by the user.

Author(s)

Vito Muggeo

References

Wood, S. N. (2001) Minimizing model fitting objectives that contain spurious local minima by bootstrap restarting. Biometrics 57, 240–244.

See Also

stepmented.lm or stepmented.glm

Examples

##See ?stepmented

[Package segmented version 2.1-0 Index]