summary.segmented.lme {segmented}	R Documentation

Summarizing model fits for segmented mixed-effects regression

Description

summary method for class segmented.lme.

Usage

## S3 method for class 'segmented.lme'
summary(object, .vcov=NULL, digits = max(3, getOption("digits") - 3), ...)

Arguments

object
Object of class "segmented.lme".

.vcov
Optional. The full covariance matrix for the parameter estimates. If provided, standard errors are computed (and displayed) according to this matrix.

digits
controls number of digits printed in output.

...
further arguments.

Details

The function summarizes and prints the most relevant information on the segmented mixed fit. The output is similar to that returned by print.summary.lme

Value

A list (similar to one returned by segmented.lm) with estimates of the variance components, and point estimates, standard errors, DF, t-value and p-value for the fixed effects. p-values for the variables U and G0 are omitted as pointless.

Author(s)

Vito M.R. Muggeo

See Also

print.segmented.lme

Examples

##continues example from segmented.lme()
# summary(os)

[Package segmented version 2.1-0 Index]