vcov.segmented.lme {segmented}	R Documentation

Variance-Covariance Matrix for a Fitted Segmented Mixed Model

Description

Returns the variance-covariance matrix of the parameters (including breakpoints) of a fitted segmented mixed model object.

Usage

## S3 method for class 'segmented.lme'
vcov(object, B=0, ret.b=FALSE, ...) 

Arguments

object
a fitted model object of class "segmented.lme", returned by segmented.lme method.

B
number of bootstrap replicates, if a bootstrap-based covariance matrix is requested.

ret.b
logical. If FALSE the full covariance matrix (for the fixed effect estimates) based on B case-resampling bootstrap samples is returned; otherwise a list with information on the bootstrap sampling distributions.

...
optional arguments, i.e. seed and it.max.b, used when implementing the bootstrap.

Details

The returned covariance matrix is based on an approximation of the nonlinear segmented term. Therefore covariances corresponding to breakpoints are reliable only in large samples and/or clear cut segmented relationships. If B>0 is set, case resampling bootstrap (on the outermost nesting level) is carried out. Moreover, if ret.b=TRUE, the bootstrap distributions are returned, rather than the covariance matrix.

Value

The full matrix of the estimated covariances of the fixed effects estimates, including the breakpoint.

Warning

All the functions for segmented mixed models (*.segmented.lme) are still at an experimental stage

Author(s)

Vito M. R. Muggeo, vito.muggeo@unipa.it

See Also

summary.segmented.lme

Examples

##continues example from segmented.lme()
# vcov(os)
# vcov(os, B=50)
# vcov(os, B=50, ret.b=TRUE)

[Package segmented version 2.1-0 Index]