vcov.stepmented {segmented}	R Documentation

Variance-Covariance Matrix for a Fitted Stepmented Model

Description

Returns the variance-covariance matrix of the parameters estimates (including breakpoints) of a fitted stepmented model object.

Usage

## S3 method for class 'stepmented'
vcov(object, k=NULL, zero.cor=TRUE, type=c("cdf", "none", "abs"), ...)

Arguments

object
a fitted model object of class "stepmented", returned by any stepmented method

k
The power of n for the smooth approximation. Simulation evidence suggests k in [-1, -1/2]; with k=-1/2 providing somewhat 'conservative' standard errors especially at small sample sizes. In general, the larger k, the smaller n^{-k}, and the smaller the jumpoint standard error.

zero.cor
If TRUE, the covariances between the jumpoints and the remaining linear coefficients are set to zero (as theory states).

type
How the covariance matrix should be computed. If "none", the usual asymptotic covariance matrix for the linear coefficients only (under homoskedasticity and assuming known the jumpoints) is returned; if "cdf", the standard normal cdf is used to approximate the indicator function (see details); "abs" is yet another approximation (currently unimplemented).

...
additional arguments.

Details

The full covariance matrix is based on the smooth approximation

I(x>ψ)\approx Φ((x-ψ)/n^{k})

via the sandwich formula using the empirical information matrix and assuming x \in [0,1]. Φ(\cdot) is the standard Normal cdf, and k is the argument k. When k=NULL (default), it is computed via

k=-(0.6 + 0.5 \ \log(snr)/√ snr - (|\hatψ-0.5|/n)^{1/2})

where snr is the signal-to-noise ratio corresponding to the estimated changepoint \hatψ (in the range (0,1)). The above formula comes from extensive simulation studies under different scenarios: Seo and Linton (2007) discuss using the normal cdf to smooth out the indicator function by suggesting \log(n)/n^{1/2} as bandwidth; we found such suggestion does not perform well in practice.

Value

The full matrix of the estimated covariances between the parameter estimates, including the breakpoints.

Warning

The function, including the value of k, must be considered at preliminary stage. Currently the value of k appears to overestimate slightly the true \hatψ variability.

Note

If the fit object has been called by stepmented(.., var.psi=TRUE), then vcov.stepmented will return object$vcov, unless the power k differs from -2/3.

Author(s)

Vito Muggeo

References

Seo MH, Linton O (2007) A smoothed least squares estimator for threshold regression models, J of Econometrics, 141: 704-735

See Also

stepmented

Examples

##see ?stepmented

[Package segmented version 2.1-0 Index]